+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/EnchantmentHelperMixin.java
@@ -7,7 +7,7 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
 
-@Mixin(value = EnchantmentHelper.class, priority = 999)
+@Mixin(value = EnchantmentHelper.class)
 public class EnchantmentHelperMixin {
     @Redirect(method = "getEnchantmentCost", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/item/ItemStack;getEnchantmentValue()I"))
     private static int onGetEnchantmentCost_getEnchantmentValue(ItemStack stack) {