+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/plugin/ModLoadedPlugin.java
@@ -19,7 +19,7 @@ public class ModLoadedPlugin implements IMixinConfigPlugin {
             "insane96mcp.iguanatweaksreborn.mixin.PistonBaseBlockMixin", () -> LoadingModList.get().getModFileById("quark") == null,
             "insane96mcp.iguanatweaksreborn.mixin.FallingBlockEntityMixin", () -> LoadingModList.get().getModFileById("quark") == null,
             "insane96mcp.iguanatweaksreborn.mixin.BoatMixin", () -> LoadingModList.get().getModFileById("quark") == null,
-            "insane96mcp.iguanatweaksreborn.mixin.EnchantmentHelperMixin", () -> LoadingModList.get().getModFileById("quark") == null
+            "insane96mcp.iguanatweaksreborn.mixin.EnchantmentHelperMixin", () -> LoadingModList.get().getModFileById("apotheosis") == null
     );
     @Override
     public void onLoad(String mixinPackage) {