+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Upcoming
 * Fixed instant heal foods leaving behind a never ending arrow
+* Fixed 'Cap to health' config option not working
 
 ## 3.2.0
 * Added Chicken from egg chance