+++ changelog.md
@@ -1,6 +1,15 @@
 # Changelog
 
+* Regenerating absorption doesn't update when it's 0
+* Border change seems to not work
+* Durability 0 on tools that break
+* Coal it's waaay too rare
+* Torch arrows place torches on moss
+
 ## Upcoming
+* Added `iguanatweaks:deathGrave` game rule
+  * If enabled the grave will generate on death
+  * Removed the config option to disable graves
 * Fixed instant heal foods leaving behind a never ending arrow
 * Fixed 'Cap to health' config option not working