+++ changelog.md
@@ -10,6 +10,8 @@
 * Added `iguanatweaks:deathGrave` game rule
   * If enabled the grave will generate on death
   * Removed the config option to disable graves
+* Fixed crash with Aphotheosis
+  * When apotheosis is present changing enchantment value of items from Item Stats will not be possible
 * Fixed instant heal foods leaving behind a never ending arrow
 * Fixed 'Cap to health' config option not working