+++ changelog.md
@@ -1,12 +1,10 @@
 # Changelog
 
-* Coal it's waaay too rare
-* Torch arrows place torches on moss
-
 ## Upcoming
 * Added `iguanatweaks:deathGrave` game rule
   * If enabled the grave will generate on death
   * Removed the config option to disable graves
+* Coal veins are much much bigger now
 * Fixed crash with Aphotheosis
   * When apotheosis is present changing enchantment value of items from Item Stats will not be possible
 * World border height is now capped to 128 blocks